rename global-helper to fs-helper
