+++ libs/ddd-angular/src/helpers/fs-helper.ts
