+++ libs/ddd-angular/src/helpers/index.ts
@@ -1,2 +1,2 @@
 export * from './domain-helper';
-export * from './global-helper';
+export * from './fs-helper';
